[llm_] Move the K8s tools to core/dsl/utils/k8s
